function_app/requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.20.0
azure-cosmos==4.5.1
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
Pillow==10.2.0
requests==2.31.0
